divided map.tsx to smaller components
