+++ src/components/Map.tsx
@@ -1,38 +1,28 @@
-import { useState, useEffect, Fragment } from 'react';
-import L from 'leaflet';
+import { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import MarkerClusterGroup from './MarkerClusterGroup';
-import BikeMarker from './BikeMarker';
 import 'leaflet/dist/leaflet.css';
 import 'leaflet-draw/dist/leaflet.draw.css';
 import mapsModel from '../models/mapModels';
 import mapModule from '../modules/mapModule';
 import Navbar from './Navbar';
+import Geofence from './Geofence';
+import Stations from './Stations';
+import Bikes from './Bikes';
+import DrawGeofence from './DrawGeofence';
 import './Map.css';
 
-import {
-  MapContainer,
-  Marker,
-  Popup,
-  TileLayer,
-  FeatureGroup,
-  Polygon,
-  useMapEvents,
-} from 'react-leaflet';
+import { MapContainer, TileLayer, useMapEvents } from 'react-leaflet';
 
-import { EditControl } from 'react-leaflet-draw';
+import ActiveBikes from './ActiveBikes';
 
 function Map() {
   const [bikes, setBikes] = useState<Array<any>>([]);
   const [stations, setStations] = useState<Array<any>>([]);
-  const [city, setCity] = useState<string>('');
+  const [, setCity] = useState<string>('');
   const [longitude, setLongitude] = useState<number>();
   const [latitude, setLatitude] = useState<number>();
-  const [mapLayers, setMapLayers] = useState<Array<any>>([]);
   const [geofence, setGeofence] = useState<Array<any>>([]);
-  const [mapBounds, setMapBounds] = useState<Array<any>>([]);
-
-  const redOption = { color: 'red' };
+  const [, setMapBounds] = useState<Array<any>>([]);
 
   /**
    * fetch bikes from API
@@ -98,85 +88,6 @@ function Map() {
   }
 
   /**
-   * Check which icon bike should have on map depending on status
-   * @param {any} scooter Current bike
-   * @returns {L.Icon<L.IconOptions> | undefined}
-   */
-  function checkIcon(scooter: any): L.Icon<L.IconOptions> | undefined {
-    const scooterIcon = mapModule.sIcon(scooter);
-    return scooterIcon;
-  }
-
-  /**
-   * Set status message
-   * @param {any} scooter Current bike
-   * @returns {string}
-   */
-  function setStatus(scooter: any): string {
-    const message = mapModule.statusMessage(scooter);
-    return message;
-  }
-
-  /**
-   * Set parking icon
-   * @returns {L.Icon}
-   */
-  function parkingIcon(): L.Icon {
-    return mapModule.pIcon();
-  }
-
-  // Insert coordinates in database
-  /**
-   * Create geofence on map
-   * @param {any} e Position to create geofence
-   * @returns {void}
-   */
-  function _onCreate(e: any): void {
-    const { layerType, layer } = e;
-    if (layerType === 'polygon') {
-      const { leafletId } = layer;
-
-      setMapLayers((layers) => [...layers, { id: leafletId, latlngs: layer.getLatLngs()[0] }]);
-    }
-  }
-
-  // Update coordinates in database
-  /**
-   * Update geofence on map
-   * @param {any} e Position to create geofence
-   * @returns {void}
-   */
-  function _onEditPath(e: any): void {
-    const {
-      layers: { _layers },
-    } = e;
-
-    Object.values(_layers).map(({ leafletId, editing }: any) => {
-      setMapLayers((layers) =>
-        layers.map((l) => (l.id === leafletId ? { ...l, latlngs: { ...editing.latlngs[0] } } : l)),
-      );
-    });
-  }
-
-  // Delete coordinates in database
-  /**
-   * Delete geofence on map
-   * @param {any} e Position to create geofence
-   * @returns {void}
-   */
-  function _onDeleted(e: any): void {
-    const {
-      layers: { _layers },
-    } = e;
-
-    Object.values(_layers).map(({ leafletId }: any) => {
-      setMapLayers((layers) =>
-        layers.filter((layers) => layers.filter((l: any) => l.id !== leafletId)),
-      );
-    });
-  }
-
-  /**
    * Resets city
    * @returns {void}
    */
@@ -187,13 +98,8 @@ function Map() {
   }
 
   /** @type {Array} filter bikes depending on status */
-  const LundBikes: Array<any> = bikes.filter(
+  const filteredBikes: Array<any> = bikes.filter(
     (bike: any) => bike.id < 1500 && bike.Status !== 40 && bike.Status !== 20,
-  );
-
-  /** @type {Array} filter bikes depending on status */
-  const StockholmBikes: Array<any> = bikes.filter(
-    (bike: any) => bike.id > 1500 && bike.Status !== 40 && bike.Status !== 20,
   );
 
   /** @type {Array} filter bikes depending on status */
@@ -242,82 +148,12 @@ function Map() {
                 attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
                 url='https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
               />
-              <MarkerClusterGroup>
-                {LundBikes.map((location: any) => (
-                  <Fragment key={location.id}>
-                    <Marker
-                      key={location.id}
-                      position={location.Position.split(',')}
-                      icon={checkIcon(location)}
-                    >
-                      <Popup key={location.id}>
-                        ID: {location.id} <br />
-                        Status: {setStatus(location)} <br />
-                        Battery: {location.Battery}% <br />
-                        <a href='#'>Move bike</a>
-                      </Popup>
-                    </Marker>
-                  </Fragment>
-                ))}
-              </MarkerClusterGroup>
-              {activeBikes.map((bike: any) => (
-                <Fragment key={bike.id}>
-                  <BikeMarker data={bike ?? {}}>
-                    {' '}
-                    <Popup key={bike.id}>
-                      ID: {bike.id} <br />
-                      Status: {setStatus(bike)} <br />
-                      Battery: {bike.Battery}% <br />
-                      <a href='#'>Stop bike</a>
-                    </Popup>
-                  </BikeMarker>
-                </Fragment>
-              ))}
-              <MarkerClusterGroup>
-                {filteredStations.map((station: any) => (
-                  <Fragment key={station.id}>
-                    <Marker
-                      key={station.station_id}
-                      position={station.Position.split(',')}
-                      icon={parkingIcon()}
-                    >
-                      <Popup key={station.id}>
-                        {station.Name} <br />
-                        ID: {station.id} <br />
-                        Occupied spots: {station.Occupied} <br />
-                        Available spots: {station.Available} <br />
-                      </Popup>
-                    </Marker>
-                  </Fragment>
-                ))}
-              </MarkerClusterGroup>
-              <FeatureGroup>
-                <EditControl
-                  position='topright'
-                  onEdited={_onEditPath}
-                  onCreated={_onCreate}
-                  onDeleted={_onDeleted}
-                  draw={{
-                    rectangle: false,
-                    circle: false,
-                    polyline: false,
-                    circlemarker: false,
-                    marker: false,
-                  }}
-                />
-              </FeatureGroup>{' '}
-              {geofence.map(
-                (location: any) => (
-                  console.log([location.Coordinates]),
-                  (
-                    <Polygon pathOptions={redOption} positions={[JSON.parse(location.Coordinates)]}>
-                      <Popup>No parking</Popup>
-                    </Polygon>
-                  )
-                ),
-              )}
+              <Bikes filteredBikes={filteredBikes} />
+              <ActiveBikes activeBikes={activeBikes} />
+              <Stations filteredStations={filteredStations} />
+              <DrawGeofence />
+              <Geofence geofence={geofence} />
             </MapContainer>
-            <pre className='text-left'>{JSON.stringify(mapLayers)}</pre>
           </div>
         </div>
       ) : (
